feat: warn about nested groups
